Added new tag generation, file uploading and request handling
Tags are now generated using an array instead of loads of parameters
making it much more flexible for you to customise your tags. You can
now also upload files with the upload method. And if the action is on
another page there is a method which handles the logic for you.

Documentation on the website to come at a later date!

v0.1.0